Refactor import/export statement
