+++ src/parser/classes/index.ts
@@ -1,0 +1,7 @@
+export { default as BinaryOperator } from './BinaryOperator';
+export { default as DataType } from './DataType';
+export { default as Parameter } from './Parameter';
+export { default as Scope } from './Scope';
+export { default as ScopeVariable } from './Scope.Variable';
+export { default as ScopeFunctionObject } from './Scope.FunctionObject';
+export { default as TokenTracker } from './TokenTracker';
