+++ src/parser/index.ts
@@ -1,6 +1,5 @@
 import * as T from "../types";
-import TokenTracker from './classes/TokenTracker';
-import Scope from './classes/Scope';
+import { TokenTracker, Scope } from './classes';
 import { parseIdentifier } from './identifier';
 import { parsePrimitive } from './primitive-literals';
 import { parseTypeLiteral } from './type-literal';
