+++ src/samples/builtin-types/list/with-function-invocation.ts
@@ -1,6 +1,6 @@
 import { Token, AST, Operator as Op } from '../../../types';
 import { NumberLiteral, Arithmetic } from '../../helper';
-import DT from '../../../parser/classes/DataType';
+import { DataType as DT } from '../../../parser/classes';
 
 const program = `\
 def addition(x: Num, y: Num): Num => x + y
