+++ src/samples/method-invocation/as-params.ts
@@ -1,6 +1,6 @@
 import { Token, AST } from '../../types';
 import { StringLiteral } from '../helper';
-import DT from '../../parser/classes/DataType';
+import { DataType as DT } from '../../parser/classes';
 
 const program = `\
 "Hello world".concat("Wyrd-Lang".concat(" is awesome!"))
